Add authentication to all notes endpoints and update front submodule

+++ api/utils.py
@@ -1,31 +1,39 @@
 from rest_framework.response import Response
-from .models import Note
+from rest_framework.exceptions import AuthenticationFailed
+import jwt
+from .models import Note, User
 from .serializers import NoteSerializer
 
 
-def createNote(request):
+def createNote(request, user):
     data = request.data
     note = Note.objects.create(
-        title=data['title'], tag=data['tag'], body=data['body'],)
+        title=data['title'], tag=data['tag'], body=data['body'], user=user)
     serializer = NoteSerializer(note, many=False)
     return Response(serializer.data)
 
 
-def getAllNotes(request):
-    notes = Note.objects.all().order_by('-updated_at')
+def getAllNotes(request, user):
+    notes = Note.objects.filter(user=user).order_by('-updated_at')
     serializer = NoteSerializer(notes, many=True)
     return Response(serializer.data)
 
 
-def getNote(request, pk):
+def getNote(request, pk, user):
     note = Note.objects.get(id=pk)
+    if note.user != user:
+        raise AuthenticationFailed(
+            "This is not your note to get, nice try tho")
     serializer = NoteSerializer(note, many=False)
     return Response(serializer.data)
 
 
-def updateNote(request, pk):
+def updateNote(request, pk, user):
     data = request.data
     note = Note.objects.get(id=pk)
+    if note.user != user:
+        raise AuthenticationFailed(
+            "This is not your note to update, nice try tho")
     serializer = NoteSerializer(instance=note, data=data)
     if serializer.is_valid():
         serializer.save()
@@ -33,7 +41,23 @@ def updateNote(request, pk):
     return Response(serializer.data)
 
 
-def deleteNote(request, pk):
+def deleteNote(request, pk, user):
     note = Note.objects.get(id=pk)
+    if note.user != user:
+        raise AuthenticationFailed(
+            "This is not your note to delete, nice try tho")
     note.delete()
     return Response('Note was deleted!')
+
+
+def getAuthenticatedUser(request):
+    token = request.COOKIES.get('jwt')
+    if not token:
+        raise AuthenticationFailed("Unauthenticated!")
+    try:
+        payload = jwt.decode(token, 'secret', algorithms=['HS256'])
+    except jwt.ExpiredSignatureError:
+        raise AuthenticationFailed("Unauthenticated!")
+
+    user = User.objects.filter(id=payload['id']).first()
+    return user
